XC-220: move BLockRange
